chore: CollectionViewModel 생성 시점을 모아보기 탭에 진입했을 때로 수정
