+++ android/feature/invitation/src/main/kotlin/com/andlife/invitation/screen/guestbook/InvitationCollectionScreen.kt
@@ -66,8 +66,7 @@ fun InvitationCollectionScreen(
             onDismissRequest = onCloseStory,
             properties = DialogProperties(usePlatformDefaultWidth = false),
         ) {
-            InvitationStoryScreen(
-                uiState = uiState,
+            InvitationStoryRoute(
                 initialIndex = uiState.selectedIndex,
                 onPageChanged = onPageChanged,
                 onToggleExpand = onToggleExpand,
